requirements.txt
aiodns
aiofiles
aiohttp
aioredis
async-timeout
attrs
better_exceptions
bottle
cchardet
gevent
git+https://github.com/iipc/urlcanon.git#egg=urlcanon
git+https://github.com/webrecorder/chrome-remote-interface-py.git#egg=cripy
git+https://github.com/webrecorder/simplechrome.git#egg=simplechrome
jinja2
pyee2
redis
requests
ruamel.yaml
ujson
uvloop
